requirements.txt
tornado==4.3
supervisor==3.3.3
requests==2.10.0
redis==2.10.5
xlwt==1.1.2
dogpile.cache==0.6.2
# yum install libcurl-devel
# export PYCURL_SSL_LIBRARY=nss
# pycurl==7.43.0
python-redis-lock==3.2.0
websocket-client==0.44.0
